Utilities

Below is the documentation of several utility functions which are hard to put in some specific category.

Dagitty.node — Function

node(dag, label)

Resolve node label in DAG: by node label return underlying graph node index

Examples

julia> using Dagitty

julia> g = DAG(:A => :C, :C => :B)
DAG: {3, 2} directed simple Int64 graph with labels [:A, :B, :C])

julia> node(g, :C)
3

Dagitty.drawdag — Function

drawdag(dag; layout=spring_layout, ...)

Draws dag with given layout method (check GraphPlot package for their description)

drawdag(dag, locs_x, locs_y)

Draws dag with given locations of every node

Examples

julia> g = DAG(:A => :C, :C => :B)
DAG: {3, 2} directed simple Int64 graph with labels [:A, :B, :C])

julia> drawdag(g, [0, 0, 1], [0, 1, 1])